Corrections de bugs

+++ Fantasy - Ravennor Expedition/Fantasy - Ravennor Expedition/Assets/Scripts/Combat/SpellEffect.cs
@@ -18,7 +18,7 @@ public class SpellEffect
     public EffectType type;
     public ActionTargets possiblesTargets;
     public int value;
-    public int scaleByLevel;
+    public float scaleByLevel;
     public Dice dicesBonus;
     public EffectTrigger trigger;
